Refactor backfilling attachments job
