+++ rust/cloud-storage/email_service/Cargo.toml
@@ -11,11 +11,11 @@ path = "src/main.rs"
 
 [[bin]]
 name = "backfill_gmail_attachments"
-path = "src/bin/backfill_gmail_attachments.rs"
+path = "src/bin/backfill_gmail_attachments/backfill_gmail_attachments.rs"
 
 [[bin]]
 name = "backfill_search"
-path = "src/bin/backfill_search.rs"
+path = "src/bin/backfill_search/backfill_search.rs"
 
 [features]
 disable_attachment_upload = []
